Add files via upload
The dataset included historical climate change-driven natural disasters. The dataset included information on the classification, type, and subtype of these disasters, along with geographical and temporal details, magnitude, economic consequences (such as reconstruction costs and damage), and administrative information. 
The primary objective of this analysis is to count the health cost of climate change driven natural disasters in Uganda